Add files via upload
This code expands upon the previous version, including further comments on specific functions throughout the code, as well as an automatic loop through all lethal fraction values between 0.0 and 1.0 in 0.1 increments. On top of this, while an invidivual CSV is produced for each run, the data is also now added to a 'Master Document' which will contain all the data for each particular population size. Hopefully, this larger document will allow for easier analysis in R.